+++ core/src/main/scala/astraea/spark/rasterframes/expressions/localops/Add.scala
@@ -22,10 +22,10 @@
 package astraea.spark.rasterframes.expressions.localops
 
 import astraea.spark.rasterframes._
-import astraea.spark.rasterframes.expressions.{BinaryLocalRasterOp, ExpressionRegistration}
+import astraea.spark.rasterframes.expressions.BinaryLocalRasterOp
 import geotrellis.raster.Tile
-import org.apache.spark.sql.catalyst.expressions.{Expression, ExpressionDescription}
 import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
+import org.apache.spark.sql.catalyst.expressions.{Expression, ExpressionDescription}
 import org.apache.spark.sql.functions.lit
 import org.apache.spark.sql.{Column, TypedColumn}